Added base auth handler to remove duplication
